requirements.txt
opencv-python-headless==4.9.0.80
ultralytics==8.2.0
paddlepaddle==3.0.0
paddleocr==2.7.3
pyyaml==6.0.1
numpy==1.26.4
